Move the test code inside the example rails app
