+++ test/example/app/controllers/child_callers_controller.rb
@@ -1,0 +1,2 @@
+class ChildCallersController < CallersController
+end
