+++ test/example/test/controllers/callers_controller_test.rb
@@ -1,0 +1,127 @@
+# frozen_string_literal: true
+
+require 'test_helper'
+
+class CallersControllerTest < ActionDispatch::IntegrationTest
+
+  test "calling_from_controller" do
+    get '/callers/calling_from_controller'
+    assert_equal "Lady of the House, speaking", response.body
+  end
+
+  test "calling_from_controller_with_params" do
+    get '/callers/calling_from_controller_with_params'
+    assert_equal "David of the House, speaking", response.body
+  end
+
+  test "calling_from_controller_with_different_status_code" do
+    get '/callers/calling_from_controller_with_different_status_code'
+    assert_equal 500, response.response_code
+  end
+
+  test "calling_from_template" do
+    get '/callers/calling_from_template'
+    assert_equal "Ring, ring: Lady of the House, speaking", response.body
+  end
+
+  test "etag_is_set_for_parent_template_when_calling_from_template" do
+    get '/callers/calling_from_template'
+    expected_etag = etag_for("Ring, ring: Lady of the House, speaking")
+    assert_equal expected_etag, response.headers['ETag']
+  end
+
+  test "internal_calling" do
+    get '/callers/internal_caller'
+    assert_equal "Are you there? Yes, ma'am", response.body
+  end
+
+  test "flash" do
+    get '/callers/set_flash'
+    assert_equal 'My stoney baby', flash[:notice]
+    get '/callers/use_flash'
+    assert_equal 'My stoney baby', response.body
+    get '/callers/use_flash'
+    assert_equal 'no flash', response.body
+  end
+
+  test "component_redirect_redirects" do
+    get '/callers/calling_redirected'
+    assert_redirected_to :controller=>"callees", :action => "being_called"
+  end
+
+  test "component_multiple_redirect_redirects" do
+    # call test_component_redirect_redirects
+    get '/callers/calling_redirected'
+    assert_redirected_to :controller=>"callees", :action => "being_called"
+    # call test_internal_calling
+    get '/callers/internal_caller'
+    assert_equal "Are you there? Yes, ma'am", response.body
+  end
+
+  test "component_as_string_redirect_renders_redirected_action" do
+    get '/callers/calling_redirected_as_string'
+    assert_equal "Lady of the House, speaking", response.body
+  end
+
+  test "calling_from_controller_with_session" do
+    get '/callers/calling_from_controller_with_session'
+    assert_equal "Bernd of the House, speaking", response.body
+  end
+
+  test "child_calling_from_template" do
+    get '/child_callers/calling_from_template'
+    assert_equal "Ring, ring: Lady of the House, speaking", response.body
+  end
+
+  protected
+  def etag_for(text)
+    %("#{Digest::MD5.hexdigest(text)}")
+  end
+end
+
+#class  < ActionDispatch::IntegrationTest
+  #setup do
+    #@caller = callers(:one)
+  #end
+
+  #test "should get index" do
+    #get callers_url
+    #assert_response :success
+  #end
+
+  #test "should get new" do
+    #get new_caller_url
+    #assert_response :success
+  #end
+
+  #test "should create caller" do
+    #assert_difference('Caller.count') do
+      #post callers_url, params: { caller: {  } }
+    #end
+
+    #assert_redirected_to caller_url(Caller.last)
+  #end
+
+  #test "should show caller" do
+    #get caller_url(@caller)
+    #assert_response :success
+  #end
+
+  #test "should get edit" do
+    #get edit_caller_url(@caller)
+    #assert_response :success
+  #end
+
+  #test "should update caller" do
+    #patch caller_url(@caller), params: { caller: {  } }
+    #assert_redirected_to caller_url(@caller)
+  #end
+
+  #test "should destroy caller" do
+    #assert_difference('Caller.count', -1) do
+      #delete caller_url(@caller)
+    #end
+
+    #assert_redirected_to callers_url
+  #end
+#end
